Remove >>#registerAt:
The >>#registerAt: system is not fully implemented and as such doesn't work. TCAPNotificationHandler defines it as subclass responsibility but only TCAPDialogueTimeoutHandler calls it. On top of it, it calls a selector that does not exist (it should probably call the >>#registerAtServicePoint: one).

Now the testcase goes through calling >>#addHandler: on the endpoint that ends up calling back into the handler. There is a strong argument for doing it that way because it allows us to set a timeout _before_ the handler is in use.