Deal with Pharo7 scheduling...
The concurrency handling is bad. The timer callback should take the lock but doesn't and the one starting the timer should take it as well.

The Q787 tests start an invocation with a 0 seconds timeout and with Pharo7 the timer callback is executed before the rest of the method. Leading to first switching to "Idle" and then back to "Sent".

This can only happen because no one is taking a lock... We need to improve this in the mid-term.